# backend/internal/httpd/apispec/openapi.yaml
openapi: 3.1.0
info:
  description: Loopback-only HTTP surface served by the Go daemon. Generated from
    Go (code-first) — do not edit by hand; run `go generate ./...`.
  title: Modern Agent HTTP daemon
  version: 0.1.0-route-shell
servers:
- description: Local daemon (loopback only)
  url: http://127.0.0.1:3001
paths:
...
  /api/v1/projects/{id}/workboard/autonomous:
    patch:
      operationId: updateWorkboardAutonomous
      parameters:
      - description: Project identifier (registry key).
        in: path
        name: id
        required: true
        schema:
          description: Project identifier (registry key).
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateWorkboardAutonomousRequest'
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkboardAutonomousResponse'
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Bad Request
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Not Found
        "409":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Conflict
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Internal Server Error
        "501":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Not Implemented
      summary: Update a project's autonomous workboard settings
      tags:
      - workboard
  /api/v1/projects/{projectId}/workboard/cards:
    get:
      operationId: listWorkCards
      parameters:
      - description: Project identifier.
        in: path
        name: projectId
        required: true
        schema:
          description: Project identifier.
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListWorkCardsResponse'
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Bad Request
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Internal Server Error
        "501":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Not Implemented
      summary: List a project's durable work cards
      tags:
      - workboard
    post:
      operationId: createWorkCard
      parameters:
      - description: Project identifier.
        in: path
        name: projectId
        required: true
        schema:
          description: Project identifier.
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateWorkCardRequest'
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkCardResponse'
          description: Created
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Bad Request
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Not Found
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Internal Server Error
        "501":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Not Implemented
      summary: Create a durable work card on a project board
      tags:
      - workboard
  /api/v1/projects/{projectId}/workboard/director-status:
    get:
      operationId: getWorkboardDirectorStatus
      parameters:
      - description: Project identifier.
        in: path
        name: projectId
        required: true
        schema:
          description: Project identifier.
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DirectorStatusResponse'
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Bad Request
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Not Found
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Internal Server Error
        "501":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Not Implemented
      summary: Return the daemon's auto-dispatch status for a project
      tags:
      - workboard
  /api/v1/projects/{projectId}/workboard/dispatch:
    post:
      operationId: dispatchWorkboard
      parameters:
      - description: Project identifier.
        in: path
        name: projectId
        required: true
        schema:
          description: Project identifier.
          type: string
      responses:
        "202":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ControllersDispatchWorkboardResponse'
          description: Accepted
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Bad Request
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Not Found
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Internal Server Error
        "501":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Not Implemented
      summary: Ask the daemon to retry dispatching Todo cards for a project
      tags:
      - workboard
...
  /api/v1/workboard/cards:
    get:
      operationId: listGlobalWorkCards
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListWorkCardsResponse'
          description: OK
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Internal Server Error
        "501":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Not Implemented
      summary: List all durable work cards across projects
      tags:
      - workboard
    post:
      operationId: createGlobalWorkCard
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateWorkCardRequest'
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkCardResponse'
          description: Created
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Bad Request
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Not Found
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Internal Server Error
        "501":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Not Implemented
      summary: Create a durable work card on a global workboard
      tags:
      - workboard
  /api/v1/workboard/cards/{cardId}:
    delete:
      operationId: deleteWorkCard
      parameters:
      - description: Work card identifier.
        in: path
        name: cardId
        required: true
        schema:
          description: Work card identifier.
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeleteWorkCardResponse'
          description: OK
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Not Found
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Internal Server Error
        "501":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Not Implemented
      summary: Permanently remove a work card and its dependent history
      tags:
      - workboard
    get:
      operationId: getWorkCard
      parameters:
      - description: Work card identifier.
        in: path
        name: cardId
        required: true
        schema:
          description: Work card identifier.
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkCardResponse'
          description: OK
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Not Found
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Internal Server Error
        "501":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Not Implemented
      summary: Fetch one work card
      tags:
      - workboard
    patch:
      operationId: updateWorkCard
      parameters:
      - description: Work card identifier.
        in: path
        name: cardId
        required: true
        schema:
          description: Work card identifier.
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateWorkCardRequest'
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkCardResponse'
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Bad Request
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Not Found
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Internal Server Error
        "501":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Not Implemented
      summary: Update mutable work-card fields
      tags:
      - workboard
  /api/v1/workboard/cards/{cardId}/dispatch-failure:
    get:
      operationId: getWorkCardDispatchFailure
      parameters:
      - description: Work card identifier.
        in: path
        name: cardId
        required: true
        schema:
          description: Work card identifier.
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DispatchFailureResponse'
          description: OK
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Not Found
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Internal Server Error
        "501":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Not Implemented
      summary: Get the latest safe automatic dispatch failure for a work card
      tags:
      - workboard
  /api/v1/workboard/cards/{cardId}/move:
    post:
      operationId: moveWorkCard
      parameters:
      - description: Work card identifier.
        in: path
        name: cardId
        required: true
        schema:
          description: Work card identifier.
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MoveWorkCardRequest'
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkCardResponse'
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Bad Request
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Not Found
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Internal Server Error
        "501":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Not Implemented
      summary: Move a work card to a status and board position
      tags:
      - workboard
  /api/v1/workboard/cards/{cardId}/nudge:
    post:
      operationId: nudgeWorkCard
      parameters:
      - description: Work card identifier.
        in: path
        name: cardId
        required: true
        schema:
          description: Work card identifier.
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NudgeWorkCardRequest'
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkCardResponse'
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Bad Request
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Not Found
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Internal Server Error
        "501":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Not Implemented
      summary: Send a nudge message to a running card's linked worker session
      tags:
      - workboard
  /api/v1/workboard/cards/{cardId}/redo:
    get:
      operationId: listWorkCardRedo
      parameters:
      - description: Work card identifier.
        in: path
        name: cardId
        required: true
        schema:
          description: Work card identifier.
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListRedoCyclesResponse'
          description: OK
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Not Found
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Internal Server Error
        "501":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Not Implemented
      summary: List Redo cycles and findings for a work card
      tags:
      - workboard
  /api/v1/workboard/cards/{cardId}/retarget:
    post:
      operationId: retargetWorkCard
      parameters:
      - description: Work card identifier.
        in: path
        name: cardId
        required: true
        schema:
          description: Work card identifier.
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RetargetWorkCardRequest'
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkCardResponse'
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Bad Request
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Not Found
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Internal Server Error
        "501":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Not Implemented
      summary: Retarget a running card's goal and hand off to Hermes or respawn the
        worker
      tags:
      - workboard
...
components:
  schemas:
    APIError:
      properties:
        code:
          type: string
        details:
          additionalProperties: {}
          type: object
        error:
          type: string
        message:
          type: string
        requestId:
          type: string
      required:
      - error
      - code
      - message
      type: object
...
    CreateWorkCardRequest:
      properties:
        agent:
          type: string
        codingAgent:
          type: string
        labels:
          items:
            type: string
          type: array
        notes:
          type: string
        priority:
          enum:
          - low
          - normal
          - high
          - urgent
          type: string
        projectId:
          type: string
        reviewerAgent:
          type: string
        reviewerMode:
          enum:
          - same
          - separate
          type: string
        scheduledAt:
          format: date-time
          type:
          - "null"
          - string
        sessionId:
          type: string
        status:
          enum:
          - triage
          - backlog
          - todo
          - scheduled
          - ready
          - running
          - review
          - testing
          - redo
          - blocked
          - done
          type: string
        targetPath:
          type: string
        testingAgent:
          type: string
        title:
          type: string
      required:
      - title
      type: object
...
    DirectorStatusResponse:
      properties:
        daemonReady:
          type: boolean
        lastDispatchAttempt:
          $ref: '#/components/schemas/DispatchAttemptResponse'
        projectId:
          type: string
        runningCount:
          type: integer
        todoCount:
          type: integer
        wipLimit:
          type: integer
      required:
      - projectId
      - daemonReady
      - runningCount
      - wipLimit
      - todoCount
      type: object
...
    ListWorkCardsResponse:
      properties:
        cards:
          items:
            $ref: '#/components/schemas/WorkCardResponse'
          type: array
      required:
      - cards
      type: object
...
    UpdateWorkboardAutonomousRequest:
      properties:
        enabled:
          type:
          - "null"
          - boolean
        mode:
          enum:
          - skip_timeout
          - short_timeout
          type:
          - "null"
          - string
        shortTimeoutMinutes:
          maximum: 1440
          minimum: 1
          type:
          - "null"
          - integer
        sticky:
          type:
          - "null"
          - boolean
      type: object
...
    WorkboardAutonomousResponse:
      properties:
        autonomous:
          $ref: '#/components/schemas/WorkboardAutonomousConfig'
        projectId:
          type: string
      required:
      - projectId
      - autonomous
      type: object
...